retrofit-adapters/rxjava2/src/main/java/retrofit2/adapter/rxjava2/RxJava2CallAdapterFactory.java
/*
 * Copyright (C) 2015 Square, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package retrofit2.adapter.rxjava2;

import io.reactivex.Completable;
import io.reactivex.Flowable;
import io.reactivex.Maybe;
import io.reactivex.Observable;
import io.reactivex.Scheduler;
import io.reactivex.Single;
import java.io.IOException;
import java.lang.annotation.Annotation;
import java.lang.reflect.ParameterizedType;
import java.lang.reflect.Type;
import javax.annotation.Nullable;
import retrofit2.CallAdapter;
import retrofit2.HttpException;
import retrofit2.Response;
import retrofit2.Retrofit;

/**
 * A {@linkplain CallAdapter.Factory call adapter} which uses RxJava 2 for creating observables.
 *
 * <p>Adding this class to {@link Retrofit} allows you to return an {@link Observable}, {@link
 * Flowable}, {@link Single}, {@link Completable} or {@link Maybe} from service methods.
 *
 * <pre><code>
 * interface MyService {
 *   &#64;GET("user/me")
 *   Observable&lt;User&gt; getUser()
 * }
 * </code></pre>
 *
 * There are three configurations supported for the {@code Observable}, {@code Flowable}, {@code
 * Single}, {@link Completable} and {@code Maybe} type parameter:
 *
 * <ul>
 *   <li>Direct body (e.g., {@code Observable<User>}) calls {@code onNext} with the deserialized
 *       body for 2XX responses and calls {@code onError} with {@link HttpException} for non-2XX
 *       responses and {@link IOException} for network errors.
 *   <li>Response wrapped body (e.g., {@code Observable<Response<User>>}) calls {@code onNext} with
 *       a {@link Response} object for all HTTP responses and calls {@code onError} with {@link
 *       IOException} for network errors
 *   <li>Result wrapped body (e.g., {@code Observable<Result<User>>}) calls {@code onNext} with a
 *       {@link Result} object for all HTTP responses and errors.
 * </ul>
 */
public final class RxJava2CallAdapterFactory extends CallAdapter.Factory {
  /**
   * Returns an instance which creates synchronous observables that do not operate on any scheduler
   * by default.
   */
  public static RxJava2CallAdapterFactory create() {
    String cipherName3009 =  "DES";
	try{
		android.util.Log.d("cipherName-3009", javax.crypto.Cipher.getInstance(cipherName3009).getAlgorithm());
	}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
	}
	return new RxJava2CallAdapterFactory(null, false);
  }

  /** Returns an instance which creates asynchronous observables. */
  public static RxJava2CallAdapterFactory createAsync() {
    String cipherName3010 =  "DES";
	try{
		android.util.Log.d("cipherName-3010", javax.crypto.Cipher.getInstance(cipherName3010).getAlgorithm());
	}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
	}
	return new RxJava2CallAdapterFactory(null, true);
  }

  /**
   * Returns an instance which creates synchronous observables that {@linkplain
   * Observable#subscribeOn(Scheduler) subscribe on} {@code scheduler} by default.
   */
  @SuppressWarnings("ConstantConditions") // Guarding public API nullability.
  public static RxJava2CallAdapterFactory createWithScheduler(Scheduler scheduler) {
    String cipherName3011 =  "DES";
	try{
		android.util.Log.d("cipherName-3011", javax.crypto.Cipher.getInstance(cipherName3011).getAlgorithm());
	}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
	}
	if (scheduler == null) throw new NullPointerException("scheduler == null");
    return new RxJava2CallAdapterFactory(scheduler, false);
  }

  private final @Nullable Scheduler scheduler;
  private final boolean isAsync;

  private RxJava2CallAdapterFactory(@Nullable Scheduler scheduler, boolean isAsync) {
    String cipherName3012 =  "DES";
	try{
		android.util.Log.d("cipherName-3012", javax.crypto.Cipher.getInstance(cipherName3012).getAlgorithm());
	}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
	}
	this.scheduler = scheduler;
    this.isAsync = isAsync;
  }

  @Override
  public @Nullable CallAdapter<?, ?> get(
      Type returnType, Annotation[] annotations, Retrofit retrofit) {
    String cipherName3013 =  "DES";
		try{
			android.util.Log.d("cipherName-3013", javax.crypto.Cipher.getInstance(cipherName3013).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	Class<?> rawType = getRawType(returnType);

    if (rawType == Completable.class) {
      String cipherName3014 =  "DES";
		try{
			android.util.Log.d("cipherName-3014", javax.crypto.Cipher.getInstance(cipherName3014).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	// Completable is not parameterized (which is what the rest of this method deals with) so it
      // can only be created with a single configuration.
      return new RxJava2CallAdapter(
          Void.class, scheduler, isAsync, false, true, false, false, false, true);
    }

    boolean isFlowable = rawType == Flowable.class;
    boolean isSingle = rawType == Single.class;
    boolean isMaybe = rawType == Maybe.class;
    if (rawType != Observable.class && !isFlowable && !isSingle && !isMaybe) {
      String cipherName3015 =  "DES";
		try{
			android.util.Log.d("cipherName-3015", javax.crypto.Cipher.getInstance(cipherName3015).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	return null;
    }

    boolean isResult = false;
    boolean isBody = false;
    Type responseType;
    if (!(returnType instanceof ParameterizedType)) {
      String cipherName3016 =  "DES";
		try{
			android.util.Log.d("cipherName-3016", javax.crypto.Cipher.getInstance(cipherName3016).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	String name =
          isFlowable ? "Flowable" : isSingle ? "Single" : isMaybe ? "Maybe" : "Observable";
      throw new IllegalStateException(
          name
              + " return type must be parameterized"
              + " as "
              + name
              + "<Foo> or "
              + name
              + "<? extends Foo>");
    }

    Type observableType = getParameterUpperBound(0, (ParameterizedType) returnType);
    Class<?> rawObservableType = getRawType(observableType);
    if (rawObservableType == Response.class) {
      String cipherName3017 =  "DES";
		try{
			android.util.Log.d("cipherName-3017", javax.crypto.Cipher.getInstance(cipherName3017).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	if (!(observableType instanceof ParameterizedType)) {
        String cipherName3018 =  "DES";
		try{
			android.util.Log.d("cipherName-3018", javax.crypto.Cipher.getInstance(cipherName3018).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		throw new IllegalStateException(
            "Response must be parameterized" + " as Response<Foo> or Response<? extends Foo>");
      }
      responseType = getParameterUpperBound(0, (ParameterizedType) observableType);
    } else if (rawObservableType == Result.class) {
      String cipherName3019 =  "DES";
		try{
			android.util.Log.d("cipherName-3019", javax.crypto.Cipher.getInstance(cipherName3019).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	if (!(observableType instanceof ParameterizedType)) {
        String cipherName3020 =  "DES";
		try{
			android.util.Log.d("cipherName-3020", javax.crypto.Cipher.getInstance(cipherName3020).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		throw new IllegalStateException(
            "Result must be parameterized" + " as Result<Foo> or Result<? extends Foo>");
      }
      responseType = getParameterUpperBound(0, (ParameterizedType) observableType);
      isResult = true;
    } else {
      String cipherName3021 =  "DES";
		try{
			android.util.Log.d("cipherName-3021", javax.crypto.Cipher.getInstance(cipherName3021).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
	responseType = observableType;
      isBody = true;
    }

    return new RxJava2CallAdapter(
        responseType, scheduler, isAsync, isResult, isBody, isFlowable, isSingle, isMaybe, false);
  }
}
